# openapi.yaml
openapi: 3.0.0
info:
  title: User Management API
  description: A REST API for managing users stored in a SQLite database.
  version: 1.0.0
  
servers:
  - url: http://127.0.0.1:5000
    description: Local development server

paths:
  /users:
    get:
      summary: Retrieve a list of users
      parameters:
        - name: search
          in: query
          description: Case-insensitive search on first or last name
          schema:
            type: string
        - name: limit
          in: query
          description: Number of items to return (default 5)
          schema:
            type: integer
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get a user by ID
      responses:
        '200':
        description: User found
      '404':
        description: User not found
    patch:
      summary: Partially update a user
      responses:
        '200':
        description: User updated
    delete:
      summary: Delete a user
      responses:
        '200':
          description: User deleted successfully

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        company_name:
          type: string
        email:
          type: string
        age:
          type: integer